refactor: add View to all view files
