+++ app/frontend/src/views/NewGameView.ts
