+++ app/frontend/src/views/StatsView.ts
